refactor(settings): move currency confirmation to dialog service
