+++ frontend/projects/webapp/src/app/feature/settings/settings.routes.ts
@@ -1,9 +1,11 @@
 import { type Routes } from '@angular/router';
 import { PAGE_TITLES } from '@core/routing';
+import { SettingsDialogService } from './settings-dialog.service';
 
 const routes: Routes = [
   {
     path: '',
+    providers: [SettingsDialogService],
     title: PAGE_TITLES.SETTINGS,
     loadComponent: () => import('./settings-page'),
   },
